refactor(analysis): use public API for config, models, and clients imports
Fix 8 internal import violations in analysis module by using public APIs:

- vibe3.config.loader.get_config → vibe3.config.get_config (2 files)
- vibe3.models.inspection.* → vibe3.models.* (1 file)
- vibe3.models.snapshot.* → vibe3.models.* (3 files)
- vibe3.clients.serena_client.* → vibe3.clients.* (2 files)

Changes:
- inspect_query_service.py:23: config import
- pr_scoring.py:16: config import
- command_analyzer.py:17: models.inspection import
- snapshot_diff.py:8-16: models.snapshot import
- snapshot_diff_section.py:7: models.snapshot import
- snapshot_service.py:12-19: models.snapshot import
- serena_file_analyzer.py:9-12: clients.serena_client import
- serena_service.py:13: clients.serena_client import

Verification:
- All 126 analysis tests pass
- No mypy type errors
- Scan shows 14 violations remain (7 exceptions + 7 dependency-blocked)

Remaining violations documented in dependency issues:
- #1702: Add CoverageReport, LayerCoverage to models public API
- #1703: Add ChangeSource types to models public API
- #1704: Add GitClient, GitHubClient to clients public API

Related: #1660, #1626
